alx_travel_app/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
celery>=5.3.0
amqp>=5.1.1